+++ src/components/sign-transactions-modal/sign-transactions-modal.types.ts
@@ -1,5 +1,5 @@
 // types here need to be synced with the types in sdk-dapp-core signTransactionsModal.types.ts
-import { EsdtEnumType, NftEnumType } from 'types/tokens.types';
+import type { EsdtEnumType, NftEnumType } from 'types/tokens.types';
 
 export interface ITransactionData {
   receiver?: string;