+++ src/components/transactions-table/components/transaction-direction-badge/tests/transaction-direction-badge.spec.ts
@@ -1,5 +1,5 @@
-// transaction-direction-badge.spec.tsx
 import { newSpecPage } from '@stencil/core/testing';
+
 import { TransactionDirectionBadge } from '../transaction-direction-badge';
 
 describe('TransactionDirectionBadge', () => {