+++ src/components/transactions-table/components/transaction-hash/transaction-hash.tsx
@@ -1,13 +1,15 @@
 import { Component, h, Prop } from '@stencil/core';
-import { ITransactionsTableRow } from '../../transactions-table.type';
+import classNames from 'classnames';
 import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
+
+import type { ITransactionsTableRow } from '../../transactions-table.type';
 
 @Component({
   tag: 'transaction-hash',
   shadow: true,
 })
 export class TransactionHash {
-  @Prop() class?: string = 'transaction-hash';
+  @Prop() class?: string;
   @Prop() transaction: ITransactionsTableRow;
 
   render() {
@@ -16,7 +18,7 @@ export class TransactionHash {
     }
 
     return (
-      <div class={this.class}>
+      <div class={classNames(this.class, 'transaction-hash')}>
         <transaction-icon iconInfo={this.transaction.iconInfo}></transaction-icon>
         <explorer-link dataTestId={DataTestIdsEnum.transactionLink} link={this.transaction.link} text={this.transaction.txHash}></explorer-link>
       </div>