+++ src/components/transactions-table/components/transaction-shards/tests/transaction-shards.spec.tsx
@@ -1,8 +1,9 @@
-import { newSpecPage } from '@stencil/core/testing';
-import { TransactionShards } from '../transaction-shards';
-import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 import { h } from '@stencil/core';
-import { ITransactionsTableRow } from '../../../transactions-table.type';
+import { newSpecPage } from '@stencil/core/testing';
+import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
+
+import type { ITransactionsTableRow } from '../../../transactions-table.type';
+import { TransactionShards } from '../transaction-shards';
 
 describe('TransactionShards', () => {
   const createMockTransaction = (senderShard: string, receiverShard: string): ITransactionsTableRow => ({
@@ -34,6 +35,7 @@ describe('TransactionShards', () => {
     },
     txHash: 'hash123',
     value: {
+      egldLabel: 'xEGLD',
       valueDecimal: '0',
       valueInteger: '100',
     },
@@ -54,11 +56,11 @@ describe('TransactionShards', () => {
         <mock:shadow-root>
           <div class="transaction-shards">
             <explorer-link class="transactions-table-body-cell-link" data-testid="${DataTestIdsEnum.shardFromLink}" link="/blocks?shard=0">
-              <span data-testid="${DataTestIdsEnum.senderShard}">0</span>
+              <span data-testid="${DataTestIdsEnum.senderShard}" slot="content">0</span>
             </explorer-link>
             <span class="transaction-shards-arrow">&#10132;</span>
             <explorer-link class="transactions-table-body-cell-link" data-testid="${DataTestIdsEnum.shardToLink}" link="/blocks?shard=1">
-              <span data-testid="${DataTestIdsEnum.receiverShard}">1</span>
+              <span data-testid="${DataTestIdsEnum.receiverShard}" slot="content">1</span>
             </explorer-link>
           </div>
         </mock:shadow-root>
@@ -78,13 +80,13 @@ describe('TransactionShards', () => {
     expect(page.root).toEqualHtml(`
       <transaction-shards class="custom-class">
         <mock:shadow-root>
-          <div class="custom-class">
+          <div class="custom-class transaction-shards">
             <explorer-link class="transactions-table-body-cell-link" data-testid="${DataTestIdsEnum.shardFromLink}" link="/blocks?shard=0">
-              <span data-testid="${DataTestIdsEnum.senderShard}">0</span>
+              <span data-testid="${DataTestIdsEnum.senderShard}" slot="content">0</span>
             </explorer-link>
             <span class="transaction-shards-arrow">&#10132;</span>
             <explorer-link class="transactions-table-body-cell-link" data-testid="${DataTestIdsEnum.shardToLink}" link="/blocks?shard=1">
-              <span data-testid="${DataTestIdsEnum.receiverShard}">1</span>
+              <span data-testid="${DataTestIdsEnum.receiverShard}" slot="content">1</span>
             </explorer-link>
           </div>
         </mock:shadow-root>